+++ LinqSamples.Introduction/Program.cs
@@ -1,5 +1,4 @@
 ﻿using LinqSamples.Resources;
-using System;
 
 namespace LinqSamples.Introduction
 {
@@ -7,11 +6,7 @@ namespace LinqSamples.Introduction
     {
         static void Main(string[] args)
         {
-            string path = @"C:\Windows";
-            Console.WriteLine("****Show large files without linq****");
-            IntroductionBase.ShowLargeFilesWithoutLinq(path);
-            Console.WriteLine("****Show large files with linq****");
-            IntroductionBase.ShowLargeFilesWithLinq(path);
+            IntroductionBase.ShowLargeFiles();
         }
     }
 }
