+++ LinqSamples.Resources/QueriesBase.cs
@@ -1,5 +1,4 @@
-﻿using LinqSamples.Resources;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
